backend/requirements.txt
python-dotenv==0.21.1
fastapi==0.92.0
uvicorn
google-cloud-storage
google-cloud-pubsub
